requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.44.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
